Modify change id type from i16 to i64

+++ database/postgresql/tbl_ext_server_shortcut.sql
@@ -1,7 +1,7 @@
 CREATE TABLE tbl_ext_server_shortcut(
 	id BIGINT PRIMARY KEY
 	, nm VARCHAR (100) NOT NULL
-	, ext_server_id SMALLINT NOT NULL
+	, ext_server_id BIGINT NOT NULL
 	, dir VARCHAR (1000) NOT NULL
 	, is_del SMALLINT DEFAULT(0)
 	, created_by VARCHAR (20)
